refactor: reorganize App component structure per Vite React best practices
- Move App component logic from src/app/App.tsx to src/App.tsx (main entry point)
- Update src/main.tsx to include ErrorBoundary and Redux Provider at root level
- Ensures correct component hierarchy and follows Vite React conventions
- src/app/ folder now contains only store configuration and feature modules
